requirements.txt
# Core ML
scikit-learn==1.4.2
xgboost==2.0.3
imbalanced-learn==0.12.3

# MLflow
mlflow==2.12.2

# Data
pandas==2.2.2
numpy==1.26.4

# Explainability
shap==0.45.0

# Config
PyYAML==6.0.1

# Visualisation
matplotlib==3.8.4
seaborn==0.13.2

# Testing
pytest==8.1.1
pytest-cov==5.0.0

# API
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic==2.7.1

# Dev tools
ruff==0.4.4
pip-audit